Problem: complexities in the first iteration of BPXE
I've implemented an initial prototype of BPXE in Elixir and while it's
(overall) not too bad, I have a certain level of anxiety when it comes
to ensuring there won't be unpleasant runtime crashes. This is mostly
about lack of enforced types in Elixir/Erlang (I am aware of Dialyzer,
but that's an afterthought, after all).

But leaving the anxiety aside (this can be managed), I think some of the
decisions I made might prove to be less than ideal.

In particular, I keep finding that self-driven "actor" flow elements
that are almost 1:1 mapping of the BPMN process are difficult to
coordinate, especially when they need to play together with other
elements (like events and event-based gateways), which leads to a lot of
complexities (synthesizing special-purpose gateways, message acking for
state saving purposes, etc.)

Solution: try and implement the same (but better) thing in Rust

Why Rust? I've had a feeling that combining its enforced type system and
futures I can build processes that do exactly what they are supposed by
chaining futures together towards the end. This basically ensures that
instantiating a process builds exactly what needs to happen instead of a
network of "self-concerned" actors.

I've also learned a lot about BPMN while working on the original iteration, and
I have this rule of thumb: you need at least a couple of rewrites to get
to something that's satisfactory. Code is not asset, it's a liability.
Bad, or unmanageable code is a bad liability. So let's start from
scratch and see if this gets any better.

This first commit doesn't do any execution just yet. It just gets us to
passable XML parsing that gives us typed data structures.