Problem: life-time bound BPMN schema types
This causes a number of issues with API ergonomics, as we have to
continuousy ensure that they are not outliving their XML source. This
approach is encouraged by strong-xml's examples to enable zero-copy
parsing.

Solution: use Self instead of str references

We're losing zero-copy parsing (we have to copy strings) but I think
this is a reasonable trade-off. We're not constantly parsing XML files
so the amount of copying is likely negligible. In return, we get
self-contained schema types that don't depend on XML source lifetime.

As an immediate gain, we can now hide document normalization from API
users so that there's no way it can be skipped.