Problem: find BPMN elements by ID
This is rather a tedious process to do every time we need to find an
element with a given ID, as it would require traversing all
sub-elements.

Solution: generate code for searching a naive search
and make it possible to figure out element type and downcast it.

This has been made possible with further amendments to codegen-rust.xsl
and downcast-rs crate.

The idea is that once an element is found, the searching party can
quickly figure out the element name (`element()` method) and downcast it
to an appropriate type.