Problem: access to single-field basic type elements
Currently we generate container structures for tags with basic type. For
example, we have this:

```rust
pub struct Incoming {
    #[xml(text, cdata)]
    pub content: Option<String>,
}
```

to access the content inside of a field using `Incoming` type we need to
use `.content`, which is a nuisance.

Solution: derive AsRef for the type and forward AsRef to the content
using `derive_more` crate