Problem: accessing single-field container elements
Currently we generate container structures for tags with basic type. For
example, we have this:

```rust
pub struct Incoming {
        pub content: Option<String>,
}
```

to access the content inside of a field using `Incoming` type we need to
use `.content`, which is a nuisance.

I've tried to solve this problem in
90a8cd0ff5525ead15e29a51f31a313e3e3e7bae, but I haven't tested it and it
doesn't work.

Solution: use Deref instead of AsRef

I guess I was absent-minded around that time :) Of course we want
`Deref`!