Problem: there is no actual engine yet
While we do have a type-safe BPMN model parser, it is not very useful by
itself. The goal of this project is to execute these models.

Solution: implement a rudimentary flow

While this is really far away from an actual engine, we need to start
somewhere and make some architectural decisions.

As a starter, the main idea behind executing processes was to use Rust's
futures to await for flow node results and proceed accordingly to
decisions made by them. This is somewhat different from the previous
version of BPXE (the one implemented in Elixir) where every flow node
was also responsible for enacting their decisions.

At this point, I've chosen Tokio as a comprehensive async library and
have implemented a rudimentary sequence flow from `startEvent` to
`endEvent`. Their firings can be observed through broadcasted events.

These events are not implemented up to the spec (they are not
interacting with an actual event system, which is currenly absent) but
are there to test and establish some foundation.

While working on this, I had to a utility function in bpmn::schema for
establishing sequence flows in a simpler manner (very useful for tests
or any other programmatic construction) and it prompted me to add
mutability traits to the auto-generated schema as well as derive `From`
for single-field containers.

I've also added a tool for unordered observation of broadcasted events
(currently limited for internal test use, but this may change),
`crate::test::Mailbox`.