Problem: orchestrating flow node flows
The initial code I wrote for running flow nodes would spawn a task for
every attempt to flow, but since we're working in an instance of a
process, we need to keep instances of flow nodes with their state to
dispatch tokens. So that code is not good.

Solution: orchestrate process in a single scheduler

Initially I've attempted something like a map of active flow nodes and
tried to provide flow node tasks access to it but it proved nearly
impossible and required non-trivial synchronization as flow node flows
need mutable access to other flow nodes.

So, instead of that, I decided to opt for a single task scheduler that
extends the original process task and maintains a list of named flow
nodes. This list is created at a startup (so we don't [yet] support
dynamic addition of flow nodes) and requires no synchronization to
access it mutably because it's a single process.

This approach is, I believe, quite nice because it shows the entirety of
flow node scheduling in a relatively short amount of space.

StartEvent has finally become a simplified event catcher (it only
supports explicit start for now).

EndEvent has become a simplified event thrower (it only supports `End`
event).

This change also adds `into_inner()` method for auto-generated `enum`
elements to extract an inner element (say, `RootElement::Process` =>
`Process`) and implements `DocumentElement` for these enums as well to
be able to figure out their `Element` variant.