Problem: start and event events by themselves are not useful
The amount of practical workflows we can build out of them is *very*
limited (only degenerative cases).

Solution: implement `parallelGateway`

This gateway allows forking and joining workflows.

As an added means of testing, `intermediateThrowEvent` with support for
throwing only `NoneEvent` was implemented. The idea here is that I
ideally want tests to rely less on `process::Log` and more on
event::ProcessEvent as these are externally observable and are
documented in specification. `process::Log` is rather an internal
observation tool.

I also added `crate::test::log_broadcast` helper that makes it easy to
print out any messages received on broadcast channels in a separate
task. This is useful for debugging tests.