Problem: no exclusive gateway
Exclusive gateway is one of the primary means of making workflow
decisions.

Solution: implement a basic exclusive gateway

This was a rather involved change. Firstly, I needed to bring some kind
of expression language so that sequence flow condition expressions will
be evaluated. I picked Rhai as a first easy-to-integrate/easy-to-use
option. It's enabled by default but can be disabled by exclusing `rhai`
feature.

Then, flow node scheduler in `Process` was augmented to probe
conditional expressions.

On the XML side, needed to be able to get `conditionExpression`
content. This element has other tags and for some reason, `strong-xml`
crate was not able to deal with that. Basically, it can't seem to be
able to handle an element with attributes, elements AND text. And that's
exactly what we need for `conditionExpression`. So, this means that
currently parsing `conditionCondition` does NOT work. That's kind of a
big deal if we care about parsing XML. It needs to be dealt with. I
amended the XML parsing test to track this but disabled it for the time
being.

I also added a bit of hack of a `xsi_type` attribute to match the
technique in which BPMN makes `conditionExpression` use
`tFormalExpression` instead of `tExpression` (through setting `xsi:type`
attribute to `tFormalExpression`).

I didn't add support for `tFormalExpression`'s `language` attribute for
the time being. Perhaps until the issue is figured out.

The good news is that the basic logic of the actual engine does not
depend on it and seems to work. But this XML issue does need to be
addressed.

I have a feeling I might need to write my own deserialization derivation
to support all necessary cases, including `xsi:type` dispatch.

I've tried other libraries (yaserde, quick-xml, serde-xml-rs) but
couldn't make them work for other reasons, either.

On a side note: I didn't feel like adding `content` to *all* elements,
but I wasn't able to figure out a better way to discern why
`tExpression` can have this data and other can't (using BPMN's
schema).

Then, I had to implement signal event throwing to be able to discern
different path firings (again, trying to avoid relying on process log
and drive development of other features instead). This also made me do
some work on other event throwing capabilities, but that is not tested
yet. Some of it may be useful as is. I am not sure.