Problem: XML parsing is down at the moment
As described in #1, XML parsing doesn't work for one of the crucial
elements (`conditionExpression`). That's a bit of a problem if one wants
to read and write BPMN documents.

Workaround: enable serde serialization and de-serialization

I wrote a simple test that shows how this would work with YAML. I tried
TOML but was getting `ValueAfterTable` error and it's way past midnight,
so I gave up.

This doesn't solve the XML solution but it does solve the problem of
reading and writing BPMN documents.