Problem: no event-based gateway
It's also a very important BPMN gateway as it allows workflows to depend
on matching for certain events to happen.

Solution: implement a barebones version of `eventBasedGateway`

This also involved creating an initial version of
`intermediateCatchEvent` (only with support for `NoneEvent` and
`SignalEvent` for the time being, roughly matching capabilities of
`intermediateThrowEvent` implementation).

I also introduced a new process::Log variant (`FlowNodeIncoming`) to be
able to trace incoming flows into nodes. Useful mostly in testing.