Problem: superfluous container structs for types
There are a lot of types in BPMN schema that are generated as a
single-field container for a primitive type (say, a string) and
accessing them is not great. They don't seem to serve any good purpose,
and have been annoying me for a while.

Examples would be `incomings`, `outgoings`, etc.

Solution: use `flatten_text` attribute for such sub-elements
and let strong-xml deserialize them to primitve types.

This also allowed me to remove one dependency `derive_more` (at least
for now).