Problem: parsing XML with expressions and formal expressions
Firstly, strong-xml seems to have a problem with elements that have
mixed types and allow for text:

```
---- parse_sample stdout ----
thread 'parse_sample' panicked at 'called `Result::unwrap()` on an `Err` value: ParsingError { error: UnexpectedToken { token: "Ok(Text { text: StrSpan(\"\\n    \" 980..985) })" } }', tests/bpmn_test.rs:7:49
note: run with `RUST_BACKTRACE=1` environment variable to display a backtra
```

Secondly, BPMN specification defines `conditionExpression` as
`tExpression` (an underspecified informal expression) but they are
routinely used in a different fashion to make them executable.

https://www.omg.org/cgi-bin/doc?dtc/10-06-02.pdf shows the use of
`xsi:type` to change the resolution of the type to `tFormalExpression`:

```xml
<conditionExpression xsi:type="tFormalExpression">
 ${getDataObject("TicketDataObject").status == "Open"}
 </conditionExpression>
```

Solution: introduce Expr type and roll a manual deserializer for it

Expr type is an enum that is dispatched based on presence of `xsi:type =
"tFormalExpression`. So this way we're typing expressions in a much
sounder way.

Closes #1
Closes #2