Problem: use of number 8 in some SmallVecs
This number is not explained and it's not clear if it should be
synchronized across the board when updated.

Solution: give this number a name and explain it

Well, in fact, give it *two* names: SMALL_INCOMING and SMALL_OUTGOING.

These numbers are just built-in thresholds for heap allocation for
SmallVecs that store some information with regards to incoming and
outgoing flows, respectively. 8 was pulled out of thin air as a
reasonably-sounding upper limit for most of flow nodes. I felt like over
8 flows on either side gets you into the realm of complex workflows.