Problem: rustfmt's formatting of tokio::select! body
For whatever reason, it can only do so much inside of that macro.

Solution: move out most of processing in Scheduler::run
to another method.

It's about time some massaging is done to that relatively long chunk of
the code, so I also went ahead and tried to untangle scheduler's code a
bit.