Problem: CI action keeps reuploading docs every time
This is not a sync, but rather a re-upload.

Solution: fix mtime timestamp of files to one date in future

From gsutil documentation:

```
Change Detection Algorithm

To determine if a file or object has changed, gsutil rsync first checks
whether the file modification time (mtime) of both the source and
destination is available. If mtime is available at both source and
destination, and the destination mtime is different than the source, or
if the source and destination file size differ, gsutil rsync will update
the destination.
```