Problem: activity container doesn't add value
Currently, `ActivityContainer` is nothing but a proxy to a flow node, so
it doesn't contribute anything.

Solution: implement support for `standardLoopCharacteristics`

This required another redesign of `crate::language` (it's more similar
to what it was before) so that we don't have to use boxed `Engine`
traits and be unable to change type of the expression by using
`eval::<T>`. What's even nicer, there's some simplification there now,
where `Engine` has to be parametrized over expression type. This allows
us to choose an appropriate evaluation mode (expression vs program) for
`FormalExpression` or `ScriptTask`.

Adds `timeout` function in `crate::test` to facilitate easy timeout
detection (see #9) as I got annoyed with hanging as a symptom of
test failure.

I also think that using `tokio::sync::watch` is probably better for
notifying tasks about the latest `Waker`. I've used `broadcast` in other
places as well as here in `ActivityContainer`. It worked in other
places, but in `ActivityContainer` tests were spontaneously failing and
that was happening until I forced the task to "catch up" on the
broadcast to get the very latest `Waker`. But `watch` channels are
exactly for this, the *last value*.