Problem: unknown WebAssembly build status
As of #10, it *should* be able to compile, but we don't know that for a fact.

Solution: instrument Cargo.tomls to be able to build for WebAssembly

Firstly, had to reduce the scope of features for Tokio (that's a good thing to
do, anyway) to make it able to compile for WebAssembly (networking and other
I/O is a no-go).

Secondly, had to add `cdylib` crate type.

At this point, I was not yet able to run tests, but at the very least we can
check that it *builds* and I've added a "wasm" job to GitHub Actions workflows
to check for that as well.