Problem: inability to run multiple instances of an activity
This is often useful for processing data, sequentially and in parallel.

Solution: implement basic `multiInstanceLoopCharacteristics`

It's lacking support for data input / output as well as event behavour,
but that's coming soon. This is just a basic way to run multi-instance
loops.

This change introduces the concept of data objects which are currently
essentially just Box`es holding anything that can be downcasted. The
reasoning behind that is so that we don't want to stipulate any
particular formats too early.