Problem: activity instances can't get the data
Thus multi-instance parallel loops are of limited usefulness.

Solution: pass data into and out of activities

It's not a complete implementation yet, but it allows to copy
information from collections to instances; as well as some other input
data associations, like data objects and properties (properties can be
both a source and a target). Limited output data associations work, too.

Association transformation is not yet implemented. Events aren't
implemented yet. Instance variables aren't implemented.

Moreso, I don't think the current design is optimal -- but it gives
something that works and can be improved. It's definitely getting to a
point where I'd consider being fragile so it's important to figure out
how to simplify it.

This change also fixes a few bugs in XML parsing/processing in
bpxe-bpmn-schema.