Problem: depending on StreamUnordered's internal order
Sounds weird, right? How can you depend on the order of an
unordered-something? Yet this is kind of what I did before. The code in
`ActivityContainer` relies on internal implementation details of
`StreamUnordered` which adds new streams to the head of the list.

Solution: order by sequentially given tokens instead