Problem: no CI tests for WebAssembly
Solution: ensure BPXE can run under WASI and Browser

This change also removes `wasm-executor` feature which was intended to
be used for testing BPXE code with wasm-executor, but isn't it better to
actually test it in an actual WebAssembly environment?

GitHub Actions, however, seems to hang on tests when they executed in a
browser (this has been the case before) so the CI steps for that are
disabled for now. At least we're testing WASI-flavoured build now.